simplify protocol manager logic
