--- zookeeper-server/src/main/java/org/apache/zookeeper/protocol/DefaultProtocol.java
@@ -1,45 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.zookeeper.protocol;
-
-import java.io.IOException;
-import org.apache.jute.InputArchive;
-import org.apache.zookeeper.proto.ConnectRequest;
-import org.apache.zookeeper.proto.ConnectResponse;
-
-/**
- * Default wire protocol for nightly ZooKeeper version.
- */
-public enum DefaultProtocol implements Protocol {
-    INSTANCE;
-
-    @Override
-    public ConnectRequest deserializeConnectRequest(InputArchive inputArchive) throws IOException {
-        final ConnectRequest request = new ConnectRequest();
-        request.deserialize(inputArchive, "connect");
-        return request;
-    }
-
-    @Override
-    public ConnectResponse deserializeConnectResponse(InputArchive inputArchive) throws IOException {
-        final ConnectResponse response = new ConnectResponse();
-        response.deserialize(inputArchive, "connect");
-        return response;
-    }
-}
