+++ zookeeper-server/src/main/java/org/apache/zookeeper/server/ServerCnxn.java
@@ -46,7 +46,7 @@ import org.apache.zookeeper.data.Stat;
 import org.apache.zookeeper.metrics.Counter;
 import org.apache.zookeeper.proto.ReplyHeader;
 import org.apache.zookeeper.proto.RequestHeader;
-import org.apache.zookeeper.protocol.ProtocolManager;
+import org.apache.zookeeper.compat.ProtocolManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
